refactor(api): introduce existing requests
