+++ frontend/src/api/types/common.ts
@@ -1,0 +1,4 @@
+export type TApiResponse<T> = {
+  success: boolean;
+  data: T;
+};
