+++ frontend/src/api/types/skill.ts
@@ -1,0 +1,4 @@
+import { Skill } from "@/entities/skill/model/types";
+import { TApiResponse } from "./common";
+
+export type TSkillResponse = TApiResponse<Skill[]>;
